drm: rockchip: dw_hdmi-rockchip: Disable HDMI QP flag for rk3399 platform
In function dw_hdmi_rockchip_bind the is_hdmi_qp is used to get HPD irq
for newer generation Rockchip boards such as rk3528 and rk3588.

And for some reason the flag is_hdmi_qp was initialized to true and the
HPD irq get function was being called for RK3399 platform as well, which
does not supports that.
Fix this by initializing the is_hdmi_qp to false for rk3399 platform.

Fixes following error messages on console:
[    9.823318] dwhdmi-rockchip ff940000.hdmi: IRQ index 1 not found

Change-Id: Ic33810902cb7d7b45fa89b79c4cc95b83edfc36a
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>